Qt: Info label fixes in the Profile dialog
When copying from another profile which has been renamed:
show the new profile name in the info label "Created from".

When copying from another profile which is later deleted:
append "(deleted)" to the info label to indicate that the origin
profile is not in the list.

Do not show "Renamed from" when a profile name if renamed back to
it's original name.

Change-Id: I0bf0c868c5dfd150a23b2ef887e7c70030b48d05
Reviewed-on: https://code.wireshark.org/review/32201
Petri-Dish: Stig Bjørlykke <stig@bjorlykke.org>
Tested-by: Petri Dish Buildbot
Reviewed-by: Stig Bjørlykke <stig@bjorlykke.org>