Don't use the same variable for the configured FCS type and the tap FCS type.
There's the value the user configured, which should neither be used nor
modified by the 802.15.4 TAP dissector; that dissector should just set
the FCS length variable.  It should also call the common dissector, as
most of the other top-level dissectors do.

That lets us have separate types for the "configured FCS type" and "tap
FCS type" variables; do so.

Speaking of calling the common dissector, the "non-ASK" dissector should
do so as well.  Make it so.

While we're at it, fail if there's an unknown FCS type in the tap
header.

Change-Id: Ib0de81764670302c771be3851e9717f0a8399ac6
Reviewed-on: https://code.wireshark.org/review/32277
Reviewed-by: Guy Harris <guy@alum.mit.edu>