Put a fake "FCS OK" value into the tree if we don't have FCS information.
That was what was being done before; do it in the main dissector
routine, as 1) the main dissector routine doesn't call the FCS or TI
CC24xx dissector if we don't have an FCS or TI CC24xx metadata trailer
and 2) that means we pull duplicate code out of those dissectors.

Also, those routines are only called if we have the full FCS/metadata
available, so there's no need for them to check for that.  (Arguably,
they should be called if the data is present, according to the reported
length, even if it's not available in the captured data, so we mark the
frame as having been cut off so the full data isn't available.)

Change-Id: I6be2a1f71a27bc41aea93e3c92743fc12c997c94
Reviewed-on: https://code.wireshark.org/review/32281
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 1cbcc57e0a0b0d4a27458799066e48fb49e2a0b5)
Reviewed-on: https://code.wireshark.org/review/32328