NSIS: Fix our extra installer location.
We know the native Windows library path early on in the configure
process. Use it to set the location of the Npcap and USBPcap installers
instead of trying to derive it from the GLib DLL path.

Change-Id: I02d48e115b374fb6080a7fe8017cba789254d5f1
Reviewed-on: https://code.wireshark.org/review/32475
Reviewed-by: Gerald Combs <gerald@wireshark.org>
(cherry picked from commit 3c6ff8cf882a7927cefcffb4ecd6dc8bfc3e814f)
Reviewed-on: https://code.wireshark.org/review/32476