MTP2: Optionally enable checksum verification
Add a protocol preference to specify whether or not the
MTP-2 frame check sequence is present in the capture file.

Allows Wireshark to correctly check the FCS and length indicator in
PCapNG files containing MTP-2 with the FCS present.  The same checks are
already possible for MTP-2 in ERF files, because Wireshark assumes that
ERF includes the CRC.

Change-Id: Ica7256818a3195ccd7f8349847446626b0ecb21c
Reviewed-on: https://code.wireshark.org/review/32513
Reviewed-by: Matthias Lang <matthias@corelatus.com>
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>