macOS: Add a notarization script.
According to

https://developer.apple.com/documentation/security/notarizing_your_app_before_distribution

notarization will be required in a future release of macOS. Add a script
for doing so. The process requires submitting our .dmg to Apple, waiting
for it to finish, and stapling a ticket to our .dmg.

Change-Id: I5b9c0c36cc2182fdd0baeada823aaacba7730a88
Reviewed-on: https://code.wireshark.org/review/32906
Reviewed-by: Anders Broman <a.broman58@gmail.com>
(cherry picked from commit c9ca78fc2bb96923828a57c47f8840d7bb3bc1dc)
Reviewed-on: https://code.wireshark.org/review/33436
Reviewed-by: Gerald Combs <gerald@wireshark.org>