cisco-fp-mim: Adding FabricPath FCS dissection and verification
FabricPath recalculates the FCS based on the whole packet (incl. the Ethernet
header) and overwrites the last four bytes of the packet, effectively stealing
the Ethernet FCS.

Since FabricPath FCS dissection and verification isn't available and falsely
attempted on the Ethernet layer, this commit implements the FCS treatment on
the FabricPath layer and treats the Ethernet layer explicitely as it would not
have a FCS.

It also adds a procotol option to enable FabricPath FCS validation which is
disabled by default though.

Bug: 15769
Change-Id: I382a4907bca158b549bcc8d77459b7829e60f94a
Reviewed-on: https://code.wireshark.org/review/33322
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>