Set the length of the DDP tvbuff based on the length in the header.
And, for DDP packets, set the length in the LLAP tvbuff based on the
length to which the DDP dissector set its tvbuff.

That lets padding be recognized as such, and also prevents dissectors
called from the DDP dissector from running past the end of the packet.

Report invalid lengths with expert info.

Change-Id: Icc6ed222a4e7b33463c7c0b02c954952fe21949a
Reviewed-on: https://code.wireshark.org/review/33142
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>