Clean up processing of RTMP data.
If the purported first tuple has a net of 0, it's a 3-octet version
indicator, not a tuple containing route information; the third octet is
a version number.  Display the version number and skip it before
displaying the tuples.

If the first tuple is an extended network tuple, the sixth octet is a
version number; display it as such.

Change-Id: I7ffb8b9df025dd75eb43eba24a37ce6bd26e8019
Reviewed-on: https://code.wireshark.org/review/33152
Reviewed-by: Guy Harris <guy@alum.mit.edu>